requirements.txt
Pillow>=8.1.0
pygame~=2.0.1
requests~=2.25.1
numpy